requirements.txt
tensorflow-datasets==4.2.0
tensorflow-text-nightly==2.6.0.dev20210417
tf-nightly==2.6.0.dev20210417
